Screen: Add a new imaging pipeline processing chain 'MirrorWindowBlit'.
This one is optionally used for dual-window output modes, if it is setup
and enabled. Otherwise a simple identity blit via the identity blit chain
is used.

Main use case is the PsychImaging task 'MirrorDisplayTo2ndOutputHead', to
implement support for an overlay window to the mirror window, allowing to
show some supplemental info to the experimenter on top of the mirrored stim.