PsychImaging: Implement overlay support for dual-window mirror mode.
Enhancements to 'MirrorDisplayTo2ndOutputHead' task:

Make use of the new 'MirrorWindowBlit' hookchain for the slave mirror
window. By default just blit the stimulus image, ie. do the mirroring.

A new optional 'useOverlay' flag to 'MirrorDisplayTo2ndOutputHead'
now allows to request the use of a overlay on top of the mirrored
stimulus in the mirror slave window, ie. a transparent layer that can
superimpose the mirrored stimulus image to provide additional infos
to the experimenter on their feedback mirror window. If the flag is
set, an additional internal offscreen window is created, fully transparent
by default, and its colorbuffer texture gets blitted on top of the
copied stimulus image each cycle. It shows wherever a non-zero alpha
channel value in the offscreen window is present, iow. alpha-test is
used for alpha-masking.

The new function overlaywin = PsychImaging('GetMirrorOverlayWindow', win);
returns a handle to the overlay window associated with the mirror window
for the given primary stimulus presentation onscreen win'dow. This works
exactly like the overlays implemented for CRS and VPixx hardware and
accessible via 'GetOverlayWindow'.

As another refinement, accept a 'specialFlags' parameter for the
slave window Open call, so user can define flags like kPsychGUIWindow
to decide if a non-fullscreen mirror window should be static, or
movable / resizable by the user, etc. The default is zero, for a fully
static mirror window.

Also, help text updates wrt. mirroring.