PsychImaging: Implement overlay support also for split-window mirror mode.
Same as for 'MirrorDisplayTo2ndOutputHead', but now for the single-window
'MirrorDisplayToSingleSplitWindow' task.

This works if FBO backed fast offscreen windows are enabled, which is the
default, and GL_EXT_framebuffer_blit is supported, which is the case on
pretty much all gpu's and Linux, Windows and macOS versions since more
than a decade, so pretty much a given.