Screen: Improve texture state management in PsychPipelineExecuteBlitter().
If any of the override texture parameters TEXTURE1D/2D/3D/RECT2D are used
so select special override texture bindings, then store all texture unit
state, and restore it at the end of PsychPipelineExecuteBlitter().

Then for any texture unit whose texture enable/binding state is changed
via override parameter, reset the unit to "all targets disabled and unbound",
before enabling and binding the target actually requested.

This to make sure no unintended textures are bound and no texture targets
are active which could interfere with the target+texture we actually want
to access in a blitter, e.g., by overriding the wanted texture source with
something unintended.

This is supposed to fix a bug in overlay handling for dual-window mirror
mode with simultaneous use of Vulkan display backend, but so far didn't
improve the situation at all. But it should have been a potential hazard
and cause for failure...