PsychImaging: Fix 'MirrorDisplayTo2ndOutputHead' overlay for Vulkan backend.
When used with the Vulkan display backend, the finalizedFBO[0] is of
texture target type GL_TEXTURE_2D, instead of the regular target type
GL_TEXTURE_RECTANGLE. The 'Builtin:IdentityBlit' blitters used by the
MirrorWindowBlit chain will always generate texture coordinates for
the finalizedFBO[0]'s texture target, so the FBO backing textures of
the overlaywin OffscreenWindow used for the mirror overlay must be of
the same texture target type, otherwise the blitter will generate wrong/
mismatched texture coordinates and things go sideways.

This is why it worked with the default OpenGL backend, but with Vulkan
it failed due to GL_TEXTURE_2D Vulkan vs. GL_TEXTURE_RECTANGLE overlaywin
mismatch. We need to be adaptive to this, and then overlay windows work
with Vulkan. This took forever to debug and fix.