ColorCal2()/PsychPowerMate(): Adapt to new dynamic USB low-level support on macOS.
Query if USB low-level access is supported, as this may or may not be the case on macOS,
depending on libusb-1.0.dylib availability.

PsychPowerMate() warns if a 'SetBrightness' call is attempted without low-level USB support.
ColorCal2() adapts to lack by using the IOPort serial-port fallback path normally used on
Windows, and warns that certain functions are unavailable.

-> This hopefully completes the transition of low-level PsychHID USB support on macOS
from macOS proprietary to the common shared libusb-1 backend.

Work time so far: 5 hours 30 minutes.