PsychHID: Unconfuse help texts and definitions for in- vs. out-transfers.
The language had it backwards! Iow. the code was correct, but confusing, and the help
text and naming of input and output arguments to 'USBControlTransfer' were definitely
confusing. Fix this.

When we are at it, add a new optional timeOutMSecs parameter to allow to specify
timeouts in milliseconds for control transfers, where the default is 10000 msecs = 10 secs,
and a value of 0 would cause an infinite blocking wait.