PsychHID: Refine support for synchronous USB bulk and interrupt transfers.
After a night of sleep, I conclude the old api is clumsy. Lets rework it to something more sane.

Only return one return argument, which is either the buffer with received data - and of the
exact length corresponding to the amount of truly received data for an IN transfer - or return
the number of bytes sent for an OUT transfer. Gone is two parameter, but sometimes redundant,
clumsiness or the weirdness of receiving an oversizes buffer with data trash at the end in case of
a shortened IN transfer. All data is valid and length() will tell us the buffer length naturally.

On OUT transfers, derive amount of data to send from length of outData vector, instead of having
it to specify it redundantly and consistently via length. Instead just ignore length.

On IN transfers, ignore the outData buffer completely.

-> This should make use of Bulk- and Interrupt transfers more clean, short and child-protected.
-> Also some help text updates. The text became way shorter and less convoluted, a good sign.

Unfortunately I can't change this for 'USBControlTransfer' as well, as it would break backwards
compatibility with existing user scripts. All we can do there is slightly improve the help text.