[Legacy]XOrgConfCreator: Add new workaround for RPi VideoCore4 on RPi OS 11.
RaspberryPi with VideoCore4 needs AccelMethod override to glamor, because
RaspberryPi OS 11, as of late 2022, disables glamor, and thereby any
hardware acceleration for OpenGL on VideoCore4 to conserve RAM, something
that doesn't fly with our use case. Force glamor accel on on modesetting
+ ARM systems, assuming this is a RaspberryPi VideoCore, and knowing it
won't hurt on any other ARM SoC with modesetting-ddx driver, as glamor is
on by default on such SoC's, so this is simply a redundant setting.

Note that according to a talk presented at XDC 2022, Igalia is already
working on a solution that allows VideoCore4 on RPi 1,2,3 to have glamor
acceleration and lower memory consumption, so this workaround may become
obsolete in the not too far future...