PsychVRHMD: Remove requirement/help text for 'GetEyeShiftVector'
'GetEyeShiftVector' was never used by any of our tests or demos or scripts,
and I can not think of any meaningful use case for it. It is very pointless.

Remove its help text / spec from PsychVRHMD(), so we don't have to support
it in new drivers like OpenXR anymore. We will keep the implementations in
the old PsychOculusVR, PsychOculusVR1 and PsychOpenHMDVR around, although
they were useless there as well.