PsychOpenXR: Add support for an approximation of 'DebugDisplay' mirroring.
If the BasicRequirement 'DebugDisplay' is used, then show the regular
onscreen window instead of hiding it, and blit copies of the finalizedFBO
framebuffer(s) into the onscreen window backbuffer, flipping it without
vsync.

This provides a basic debug view onscreen, which approximates what the
subject gets to see inside the VR HMD, or AR/MR display device.

Note that we show the mono image or stereo image pair that is sent to
the XR compositor, not the actual image generated by the compositor and
displayed on the XR display device, as opposed to what, e.g., the Oculus
VR1 driver does. The reason is that there doesn't exist a way to access
the final XR composited output image, ie. no api from OpenXR or from any
OpenXR extension. This is likely to stay this way, as apparently members
of the OpenXR working group were strictly against providing access to the
final composited image, not even in debug modes. This could be due to
concerns about user privacy, or about digital content protection, but it
is what it is.

Cfe. the following comment of a member of the OpenXR working group in
response to a feature request wrt. compositor image readback:
https://community.khronos.org/t/support-for-reading-back-the-compositor-output/108853/2

Note: This current implementation could benefit from some future
performance optimizations, tbd.