PsychHID: Add PsychHID('USBClaimInterface', usbHandle, interfaceId).
This function allows to explicitely claim a USB interface to enable it
for I/O from/to an USB interface endpoint. Bulk- or interrupt transfers
don't work if the interface who owns the endpoint has not been claimed.

If a call to this function is omitted before doing bulk or interrupt
transfers, then PsychHID will automatically claim interface 0.

Claimed interfaces are auto-released when closing an USB device.

Kernel drivers potentially attached to - and blocking - an interface
will be automatically detached, and then reattached as device close.

In other words: Use of interface 0 does not need any extra user code.
Use of other interfaces will require this call in time.