Cohorts/libusb1-win32: Update libusb-1 for Windows to version 1.0.26.
11 Years after the previous upgrade it is time for a little refresher.
This brings API support on par with Linux and macOS, and 11 years of bug fixes
and improvements. With this upgrade we can open various USB devices we could
not open with the old pre 1.0.9 version. Interrupt transfers now also work, so this
at least brings us into the ballpark of the Linux variant.

Zip files with Windows/macOS binaries and the corresponding source code
are included. Part of the zip files needed for building PsychHID are expanded
into the old directory structure, but unneded parts of the old directory structure
are now deleted to avoid confusion and cut it down to the essentials.