Cohorts: Add libusb-1.0.dylib for 64-Bit Intel arch to libusb-win32/
Just extracted from the binaries zip file, so also v1.0.26. Mostly, so when building PsychHID
for Python, one can weak_link against this copy to no need an external copy just for building.
We don't provide a runtime copy yet, as Python might have other installation/distribution/
dependency mechanisms for that and with weak_link'ing, this is optional at runtime.