PsychOpenXR[Core]: Refactor for proper timing.
On init, in processXREvents(), triggered by a call to 'Controller' atm.,
the transition to XR_SESSION_STATE_READY, followed by a xrBeginSession()
triggers the initial sync via PresentExecute(openxr, TRUE); for an "inInit"
execution, which calls xrWaitFrame() for initial sync-up, and then
xrBeginFrame() to allow rendering and presentation.

Every successive call to PresentExecute(openxr, FALSE); as part of
Screen('Flip') -> PsychOpenXRCore('PresentFrame', ...); triggers a
xrEndFrame() -> xrWaitFrame() -> xrBeginFrame() sequence. This way
the proper ordering of wait->begin=>end->wait->begin=>end->wait->begin...
is achieved, but also xrEndFrame() + xrWaitFrame() are now called
back-to-back for proper timing and timestamping. Or so goes the theory...

This provides the baseline implementation for testing and refining
presentation timing and timestamping.