PsychHID: Add workarounds for scripts using 'USBControlTransfer' wrong.
Control transfers with endpoints other than zero as targets, or with
interfaces as targets, must claim the interface or interface associated
with an endpoint before use. Explain that in the help text for
USBControlTransfer. In case a buggy script does not do that, the Linux
kernel at least will auto-claim the relevant interface iff no kernel
driver is attached to that interface already. This can cause wrong
interactions with manual claiming of the same auto-claimed interface,
or implicitely with interrupt/iso/bulk transfers. Lets add some logic
to detect such cases and tries to handle them, while simultaenously
educating users about the buggy script and how to fix it.

This will hopefully fix the problems with getting the ColorVision SpyderX
going on Linux.

See detailed explanation of the issue and solution in comments in the
code of this commit.

Testing with PowerMate shows good results. Test with SpyderX by user is
pending.