PsychOpenXRCore: Work around broken SteamVR/Linux xrDestroyInstance().
Only call xrDestroyInstance() if the OpenXR runtime is not Linux + SteamVR's
OpenXR, because otherwise we hang until killed in xrDestroyInstance() due to
a SteamVR bug present since at least January 2021 and unfixed by Valve as of
December 2022 - Strong work!

See: https://github.com/ValveSoftware/SteamVR-for-Linux/issues/422
for the bug report, and
https://github.com/cmbruns/pyopenxr/pull/60
for a similar workaround in pyopenxr.

Our "solution" for now is to not call xrDestroyInstance() on Linux+SteamVR,
as that would hang, but instead skip that call and instead try to pin the
openxr loader into process memory via a dlopen() call to that effect.

This lets the XR session exit normally, but still makes it impossible to
have a second run of a XR script until application restart! No simple way
around this. Also it just replaces an immediate Octave/Matlab hang by a
deferred crash when quitting the application! Still horrible, but better
than hanging at an even more inconvenient time.

Ofc. the correct solution would be for somebody fixing this damned
SteamVR bug, but apparently nobody cares. The joys of proprietary software.