PsychOpenXR: Add yet another workaround for broken SteamVR OpenXR runtime.
SteamVR OpenXR runtime needs a workaround for not properly
managing its OpenGL context sometimes. So far confirmed to be
needed on Linux with SteamVR 1.24.6. Status on Windows not yet known.

It leaves our OpenGL context in a disabled state after a call to
xrReleaseSwapchainImages(), which would cause OpenGL errors. Fix this,
because the creators of SteamVR apparently can't be bothered to fix it.