PsychOpenXR: Add workaround for SteamVR with Oculus backend on MS-Windows.
The SteamVR runtime on MS-Windows in 3D rendering mode needs fp16
buffers for some absurd reason, most likely some bug in handling of projection layers.

This workaround is not needed in 2D mode where quad layers are used.
This at least as of December 2022 with SteamVR version 1.24.7, when using
OculusVR as backend to drive Oculus devices. Not clear yet if this is also
needed for other vendors HMDs, or only when using the OculusVR runtime
for Oculus HMDs as backend for SteamVR. Go figure!