PsychOpenXR[Core]: Fix broken controller input on SteamVR/Linux.
After a whopping 20 hours+ debug nightmare, the culprit for totally
not working VR controller input (as tested with Oculus touch controllers
in a Linux + SteamVR + Monado SteamVR driver plugin config) is that our
.applicationName contained a / character for the Octave version of the
PsychOpenXRCore driver, ie. the name string "GNU/Octave 64-Bit" contains
a "/" !! SteamVR uses the application name without proper escaping of
special characters to build filenames for config files, so the / turned
into a filesystem path/directory separator on Unix systems like Linux and
config files could not be created. This lead to interaction profile binding
failure, and in turn to failure to get any controller input!

For reference, the SteamVR generated files are created under (Ubuntu):
$HOME/.steam/debian-installation/config/openxr/ with an applicationName +
host app (matlab/octave/python/...) component.

So we replace PSYCHTOOLBOX_SCRIPTING_LANGUAGE_NAME with a hard-coded name
of "Psychtoolbox". Less sensical, but fixes the problem.

This bug present at least in SteamVR 1.24.7 as of January 2023.

Bump copyright year to 2023 when we are at it.

-> This leaves the following known problems at least for Oculus touch
controllers under SteamVR / SteamXR:

- Controller type is not detected, ie. PsychVRHMD('Controllers', hmd)
  returns 0 for "no controllers detected" on both Linux and Windows with
  SteamVR.

- Grip/Squeeze buttons don't work on Windows or Linux so far.

- Thumbsticks and their click buttons do not work on Linux.