PsychOpenXRCore: Enable Linux-only controller button bindings.
These do not work on Windows with OculusVR runtime, but do work on
Linux with both Monado and SteamVR, iow. with all the Linux OpenXR
runtimes i know of. Therefore compile-time enable them on Linux.

As these are /system/click/ buttons, it is not clear if those would
only be reported on Linux+Monado, or also on Linux+SteamVR, and I
don't have suitable test hardware. But at least it could give us a
OVR.Button_ENTER if it worked, so lets at least allow for the option.

Test on Windows+SteamVR pending.