PsychOpenXRCore: Refine targetDisplayTime calculation.
If user script did not specify a target time, set it to something of at least "now", in case the
xrWaitFrame provided most recent predicted time is outdated and in the past, instead of in
the future. Mostly so we don't spec a time in the past after an animation loop was halted for
a while.

I hoped this would help SteamVR to deal better with calling a 'Flip' after some pause in the
script execution, but no dice on Windows with SteamVR + OculusVR backend.

Anyway, it didn't help here, but it should not hurt, and in any way it might help in some other cases.