PsychOpenXRCore: Implement better workaround for SteamVR/Linux xrDestroyInstance bug.
On a buggy SteamVR runtime on Linux, not just skip xrDestroyInstance(), but the complete
driver shutdown procedure, so the XrInstance can be recycled across multiple work sessions.

This is now possible due to refactoring of input initialization. It can't fix the dumb SteamVR bug,
but it should allow to keep recycling the same XrInstance throughout the lifetime of a users
Octave or Matlab work session, avoiding the need for xrDestroyInstance altogether, thereby
side-stepping the problem.

So far successfully tested on MS-Windows, actual test on Linux + SteamVR pending.