PsychOpenXRCore: Delete wrong description of how timestamps come to be.
This would have been right under a more simple XR compositor, but it is
not correct for either Monado, nor for OculusVR or SteamVR, or likely any
other runtime. Especially not if things like asynchronous timewarp, and
sync or async reprojection and similar techniques or throttling mechanisms
are (dynamically) active.

See hacks in PsychOpenXR.m with double PresentFrame for Monado, which give
at least approximately correct (optimistic, ie. potentially 1 frame too
early) timestamps under Monado. However that's not good enough, and it
incurs a performance hit of cutting fps in half.

SteamVR/OculusVR are so broken wrt. timing, this is pointless. We'll need
some special Monado specific hack and a proper implementation of a draft
of a proper OpenXR timestamping extension, as the only way forward.

Lets declutter the code until then...