PsychOpenXR[Core]: Add workarounds in multi-threaded mode for MS-Windows interop bugs.
Windows OpenXR runtimes need a special workaround in multi-threaded
mode to deal with severe limitations of MS-Windows OpenGL-DirectX
interop extension (WGL_NV_DX_interop) wrt. use of Direct3D interop
textures shared across multiple OpenGL contexts. Essentially, only one
OpenGL context at a time can lock and use such a texture, but we don't
have any control over locking, as this is hidden inside the OpenXR
runtimes (xrAcquire/Wait/ReleaseSwapchainImage implementations). With
the "OpenXR OpenGL work context on one thread only at a time" - which
is neccessarily the presenterThread in multi-threaded mode, makes it
almost impossible to also use such a OpenXR swapchain image texture as
finalizedFBO backing for Screen's stimulus rendering and post-
processing --> Resource locking disasters and other monsters will come
out! We work around this by keeping the XrSwapchain textures away from
Screen: Instead we use Screen's own backing textures for the
finalizedFBO's, only use OpenXR textures in the presenterThread, and
then inside the presenterThreads releaseImages function, we copy the
rendered stimulus from Screen's finalizedFBO texture to the OpenXR
swapchain texture -- one extra full stimulus image copy per present!
Not good for performance, but this way Screen does not have to touch
OpenXR stuff, and the OpenXR thread's bound OpenXR work OpenGL context
can read from the texture shared with Screen, and write to the OpenXR
textures. Cfe. special code in PsychOpenXRCore's releaseTextureHandles()
routine. Here we need delicate switching between the two modes of
operation.
So far the theory: In practice, this only fixes mayhem on the OculusVR
runtime, but SteamVR still shits itself when using a secondary OpenGL
userspace rendering context via Screen('Begin/EndOpenGL'). May be a
related bug or not, but this is not it yet...

As feared, use of multi-threading drives complexity and bugs through the roof! It worked
beautifully on Linux with Monado and SteamVR, but on MS-Windows we are in for a total
nightmare. Unfortunately we can't avoid multi-threading, except for Monado on Linux :(