PsychOpenXR[Core]: Rewrite the multi-threading workaround for broken Windows runtimes.
The previous approach was just too shaky. Make it simpler and sacrifice a bit of performance in
exchange. As soon as multiThreaded is non-zero, ie. there is at least the possibility that the async
background presenter thread will be used (even only as a dynamic opt-in), we always keep the
textures of Screen() separate from the textures provided by the OpenXR swapchain:

1. Screen's finalizedFBO's have their own private colorbuffer attachment textures.
2. PsychOpenXRCore will copy the content of those textures into the xrSwapchain image texture
whenever releaseTextureHandles() is called for latching a new stimulus frame for presentation.
As this happens on the OpenXR work context, that context can access both the finalizedFBO
textures shared with Screen and its OpenGL contexts (standard context resource sharing) and
the special OpenXR textures shared with DirectX/Direct3D XR compositors via WGL_NV_DX_interop
extension and all its constraints and limitations.

Additionally we need to add some manual OpenGL context switching code in PresentExecute() for
single-threaded mode, at least for the OculusVR runtime, as this runtime seems to suffer from as
many absurd OpenGL context management bugs as its SteamVR counterpart - just at different
places, to keep things "interesting".

This now works well in single-threaded, multi-threaded and dynamic ST/MT on-demand mode
with the OculusVR OpenXR runtime as of version 1.79.0 from January 2023.

SteamVR is its own shit-show though...