PsychOpenXR[Core]: Make robust against removal of HMD from users head.
Make sure a session can successfully start up while the XR device is not reported as present
on the users head (proximity sensor reports "Head missing"), that it survives removal of the
device mid-session and setting it up again, and that the session can shut down cleanly while
the HMD is not on the users head.

The procedure during multithreaded operation is a bit fishy, but seems to work well enough.