PsychOpenXR: Implement decision logic of when to en/disable multithreaded mode.
Try to encode all known limitations and caveats and peculiarities of the so far tested
OpenXR runtimes on different operating systems under different conditions and let
the driver decide if it wants to disable multithreading MT, enable it, or conditionally
enable it on-demand. Try to add warning and info output about decisions made.

Various new keywords for the basiqRequirements parameter can be used to influence the
decision making about how and when to use multi-threading or single-threading.

'[No]TimingSupport' if the script needs reasonably precise visual stimulus onset timing,
iow. if it try hard to obey tWhen target times passed via vbl = Screen('Flip', win, tWhen);
or if it can play fast and loose with those, or not satisfy them at all.

'[No]TimestampingSupport' if the script needs reasonably precise visual stimulus onset
timestamp returned in 'vbl' of Flip, or if it can return unreliable, imprecise - or outright
wrong! - timestamps, because it doesn't matter for this script.

'Use2DViewsWhen3DStopped' If the driver should switch to OpenXR quadViews, aka 2D
mode when in 3D mode and PsychVRHMD('Stop') is called. quadViews can often be kept
visually stable without need for multi-threading, the normally used projectionLayers in 3D
almost never. So using this keyword may allow to avoid multi-threading. But as the
projection geometry of the two view types is different, visual glitches/disruptions may
occur at PsychVRHMD('Stop/Start') calls, unless the location and size of the 2D views is
fine-tuned by the script for the use case to minimize visual discontinuities.

'DontCareAboutVisualGlitchesWhenStopped' Keep using projectionLayers in 3D and quad views
in 2D and simply don't try to handle Stops, accept the visual artifacts. This if one wants to avoid
multi-threading (the standard remedy against glitches) at all cost.

'ForbidMultiThreading' Forbid multi-threading no matter what, e.g., to work around broken
proprietary OpenXR runtimes like SteamVR on Windows, or to squeeze out every bit of
performance, even if this breaks timing, timestamping and causes visual artifacts, jerks, jitter,
disruption etc.

If a script does not provide the [No]Timing and [No]Timestamping keywords to clue the driver
in on the scripts timing needs, the driver will try to derive from the scripts actual use of the
Screen('Flip'. win [tWhen]) command, about its needs. If a non-default tWhen is provided, then
the driver assumes from then on that timing is important and enables all measures. Otherwise
it will assume it does not matter and stay single-threaded for higher performance. It is best to
specify the clues though, as the driver may misguess -- if only timestamps are requested, but
no tWhen target times specified, the driver won't know that timestamps matter, and timestamps
will most likely be wrong/inaccurate or at least not robust and trustworthy. Also, even being on
the fence about potential timing needs produces some extra overhead that can reduce the
overall performance by adding a millisecond or two of gpu overhead per frame.

Also update most VR demos to make use of the new functionality for an optimal timing
vs. performance tradeoff. And add new optional tests to VRInputStuffTest to exercise some
of this new stuff.

If multi-threading is used in any given mode of operation depends on the keywords passed in,
in conjunction with encoded knowledge about flaws, bugs and limitations of already tested
OpenXR runtimes. Currently SteamVR on Windows with OculusVR backend needs the  most
workarounds, followed by OculusVR on Windows and SteamVR on Linux sharing a 2nd place.
The open-source MonadoXR runtime on Linux has the least limitations and quirks - According
to the current OpenXR spec, it has none and is perfectly spec compliant. However OpenXR
itself even with a perfect runtime currently has no provisions for proper timestamping, so even
on Monado we need multi-threading workarounds for that as of this commit. I have plans for
special Monado specific hacks, and functional extensions to Monado to solve this problem and
make Monado the Nr. 1 choice for applications which require precise timing/timestamping and
low level control, but this is Tbd as of this commit.

Successfully tested on Windows + OculusVR OpenXR runtime and Windows + SteamVR
runtime with OculusVR backend for Rift CV-1 HMD so far.