PsychOpenXR[Core]: function to change 2D quad view position and size.
This is exposed via...
[oldPosition, oldSize] = PsychVRHMD('View2DParameters', hmd, eye [, position][, size]);
..., but only has a proper implementation for PsychOpenXR. All older drivers have a dummy
implementation that does nothing, goes nowhere, just to keep script running. These return
oldPosition as [NaN, NaN, NaN], oldSize as [NaN, NaN], and ignore any new settings provided
to them.

We need this to optimize size and location relative to the optical centers and looking directions
of the eyes, for 2D stereoscopic displays, but also for the 'Use2DViewsWhen3DStopped' setting
in 3D rendering mode in order to reduce visual disruption when switching between
PsychVRHMD('Start/Stop').

When we are at, adapt default 2D quad view size to a virtual height of 1 meter and a width
adapted to the aspect ratio of the stimulus framebuffer, so circles look circly and not elliptic,
and squares are squares and not rectangles, etc.

VRInputStuffTest contains some new test code for this, allowing adjustments with horizontal
movements of the right thumbstick on a right hand controller.