PsychHID/Windows: Convert DirectInput8 code from C++ to C.
I hoped this would be the troublemaker causing failure to load on Octave-8 rc1, but not dice!

Anyway, the work has been done, and C is generally much less of a headache than C++, so maybe
it helps avoid - or at least diagnose and resolve more easily - unrelated future trouble.