Python: setup.py: Update for optional use of libusb-1.0 with PsychHID on macOS.
We only weak link, and don't distribute libusb-1.0.dylib. Its actual presence on users
systems is optional, as long as users don't use PsychHID functions like
USBControlTransfer/USBInterruptTransfer/USBBulkTransfer, for which i suspect Python
already has standard packages. But if we should bundle is a the discretion of the
Python maintainers.

This is totally untested, I don't have the time to setup a pip environment right now.

@aforren1 have a look, test, send pull requests with fixes if needed asap!