PsychOculusVR: Remove for macOS. No VR on macOS as of PTB 3.0.19
It only supports the out-of-sale since many years Oculus Rift DK1 and Rift DK2,
with an Oculus v0.5 runtime for macOS that is not available for download from
Oculus or anywhere else anymore since years, and only for macOS versions which
support 32-Bit Intel architecture executables, iow. doesn't work on macOS 10.15
Catalina and later anymore.

-> Quite pointless to continue shipping this. -> Remove, but leave build-script
code in disabled fashion, just in case.

=> This means that Psychtoolbox 3.0.19 will no longer support any VR hardware on macOS.