PsychVRHMD('GetStaticRenderParameters'): Return field of view as well.
It is expressed as angles in units of radians. This is useful for PsychOpenXR to allow
sizing and positioning the 2D views optimally for a given paradigm if users are not
happy with the default setup. It provides the crucial info about field of view and
thereby shape of the view frustum.

We implement this by a documentation update in PsychVRHMD, and suitable
compatibility code for the legacy drivers.