PsychOpenXR: Do more clever default setup of 2D quad view position and size.
Try to match semantic of the legacy drivers more closely, and to try to optimize
size and head-relative position for the given field of view of the given XR HMD.

Similar to the other drivers, computing view props (wrongly) assuming a symmetric
field of view for both eyes gives a better result than doing the right thing of handling
asymmetric field of view.

Now we simulate a display screen floating 1 meter away from the viewers eyes,
centered on the viewers line of sight (for straight ahead case) / optical axis.

One limitation of mono mode seems to be that xrLocateViews does not work, so
we can't get FoV for mono quad view, and have to fudge it with hard-coded values
taken from the Oculus Rift CV-1 at default settings. At least this failure was present
with the proprietary OculusXR runtime on Windows.