VRInputStuffTest: Add optional test for reference space selection on OpenXR.
Test code is disabled, but would switch to STAGE reference space for tracking.
The STAGE space shifts the origin to the ground floor level on Oculus OpenXR
or SteamVR OpenXR, and returns the bounding rectangle of the tracking area,
whereas the default LOCAL space doesn't, and has origin height more at eye
height in a sitting position of the subject.