PsychOpenXR: Allow userscript to override field of view for calculations in mono mode.
Now user-script provided field of view in 'SetupRenderingParameters' always takes
precedence. If absent, the OpenXR driver will be queried for the HMD's actual fov,
and if that fails, it will fall back to hard-coded fov of the Oculus Rift CV-1.

Also add some debug message, so user knows what happens.

Note: At least with the proprietary OculusVR1 OpenXR runtime for MS-Windows,
xrLocateViews does not work for monoscopic mode, so we can't query the driver
for actual fov in that mode and have to fallback to the hard-coded Oculus Rift CV-1
values. Not sure if this is per spec or a runtime bug/limitation.

-> This now allows users to choose arbitrary 2D view sizes by specifying arbitrary
override field of views.