PsychVRHMD/PsychOpenXR: Add new basicRequirement  '2DViewDistMeters='
It allows to position the virtual viewscreen in 2D mode at a different override distance
than the default 1 meters, for which quadView (x,y) position and size is calculated to
fill the field of view. This allows to fit more of the stimulus image into the FoV, or less,
depending on the needs of the paradigm.

E.g., '2DViewDistMeters=2.1' would place the quadView 2.1 meters away, instead of
the default 1.0 meters.