Build scripts: Enable PsychOpenXRCore build for Linux and Windows.
Enable build for 64-Bit Octave and Matlab on Linux.
Only enable build for 64-Bit Matlab on Windows, not the Octave build.

The statically built openxr_loader.lib for 64-Bit Intel on Windows is built with
MSVC 2019 and incompatible with the MinGW64 setup for Octave.

I've wasted enough unpaid time with this stuff so I won't bother fixing this
problem for Octave on Windows now, until more funding becomes available.

So for now: OpenXR support for 64-Bit Matlab on Windows only.